ccl4data {deSolve}	R Documentation

Closed Chamber Study of CCl4 Metabolism by Rats.

Description

The results of a closed chamber experiment to determine metabolic parameters for CCl4 (carbon tetrachloride) in rats.

Usage

data(ccl4data)

Format

This data frame contains the following columns:

time

the time (in hours after starting the experiment).

initconc

initial chamber concentration (ppm).

animal

this is a repeated measures design; this variable indicates which animal the observation pertains to.

ChamberConc

chamber concentration at time, in ppm.

Source

Evans, et al. 1994 Applications of sensitivity analysis to a physiologically based pharmacokinetic model for carbon tetrachloride in rats. Toxicology and Applied Pharmacology 128: 36 – 44.

Examples

plot(ChamberConc ~ time, data = ccl4data, xlab = "Time (hours)",
  xlim = range(c(0, ccl4data$time)),
  ylab = "Chamber Concentration (ppm)", log = "y")
ccl4data.avg <- aggregate(ccl4data$ChamberConc,
                  by = ccl4data[c("time", "initconc")], mean)
points(x ~ time, data = ccl4data.avg, pch = 16)                        

[Package deSolve version 1.28 Index]