diagnostics {deSolve}	R Documentation

Print Diagnostic Characteristics of Solvers

Description

Prints several diagnostics of the simulation to the screen, e.g. number of steps taken, the last step size, ...

Usage

diagnostics(obj, ...)
## Default S3 method:
diagnostics(obj, ...)

Arguments

obj
is an output data structure produced by one of the solver routines.

...
optional arguments allowing to extend diagnostics as a generic function.

Details

Detailed information obout the success of a simulation is printed, if a diagnostics function exists for a specific solver routine. A warning is printed, if no class-specific diagnostics exists.

Please consult the class-specific help page for details.

See Also

diagnostics.deSolve for diagnostics of differential equaton solvers.

[Package deSolve version 1.35 Index]